+++ examples/simple_model.cc
@@ -79,7 +79,7 @@ void buildNextStateFunction(const char* const* events, int nEvents,
     maxBound = MAX(maxBound, d->getVariableBound(i, false));
   }
   maxBound++;
-  int* temp = new int[maxBound];
+  long* temp = new long[maxBound];
   int* minterm = new int[nVars+1];
   int* mtprime = new int[nVars+1];
   dd_edge** varP  = new dd_edge*[nVars+1];
